move bridge out of the experimental directory
