+++ bridge/opentracing/internal/doc.go
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal // import "go.opentelemetry.io/experimental/bridge/opentracing/internal"
+package internal // import "go.opentelemetry.io/bridge/opentracing/internal"
