+++ bridge/opentracing/migration/api.go
@@ -15,7 +15,7 @@
 // This package provides interfaces and functions that are useful for
 // providing a cooperation of the OpenTelemetry tracers with the
 // OpenTracing API.
-package migration // import "go.opentelemetry.io/experimental/bridge/opentracing/migration"
+package migration // import "go.opentelemetry.io/bridge/opentracing/migration"
 
 import (
 	"context"
